requirements.txt
pydantic
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
pydantic-settings
jwt
authlib
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
deepface
scikit-learn
Pillow
numpy